requirements.txt
asgiref==3.7.2
coverage==7.3.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-import-export==3.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
iniconfig==2.0.0
MarkupPy==1.14
model-bakery==1.15.0
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pluggy==1.3.0
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
tomli==2.0.1
typing_extensions==4.7.1
xlrd==2.0.1
xlwt==1.3.0
